+++ lib/fluent_result.dart
@@ -3,4 +3,5 @@ export 'src/extensions.dart';
 export 'src/methods.dart';
 export 'src/result.dart';
 export 'src/result_config.dart';
+export 'src/result_matcher.dart';
 export 'src/result_of.dart';